requirements.txt
accelerate==0.18.0
agilerl==0.1.30
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
cachetools==5.3.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
distinctipy==1.3.4
distlib==0.3.8
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
fastrand==1.8.0
filelock==3.14.0
flatten-dict==0.4.2
fonttools==4.53.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gymnasium==0.28.1
h5py==3.11.0
huggingface-hub==0.23.2
hydra-core==1.3.2
identify==2.5.36
idna==3.7
imageio==2.34.1
jax-jumpy==1.0.0
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
minari==0.4.3
mpmath==1.3.0
networkx==3.3
nodeenv==1.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.0
pathtools==0.1.2
pettingzoo==1.24.3
pillow==10.3.0
platformdirs==4.2.2
portion==2.4.2
pre-commit==3.7.1
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==4.6.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
SuperSuit==3.9.2
sympy==1.12.1
termcolor==1.1.0
tinyscaler==1.2.7
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
triton==2.3.0
typer==0.12.3
typing_extensions==4.12.0
urllib3==2.2.1
virtualenv==20.26.2
wandb==0.13.11
